+++ src/Chat/index.js
@@ -1,67 +1,68 @@
-import React, { useEffect, useState } from "react";
-import MessageList from "./MessageList";
-import MessageInput from "./MessageInput";
-import FolderList from "./FolderList";
-import { makeStyles } from "@material-ui/core/styles";
+import React from "react";
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  Link,
+  useRouteMatch,
+  useParams
+} from "react-router-dom";
 
-const useStyles = makeStyles((theme) => ({
-  chatWrapper: {
-    height: "100%",
-    width: "100%",
-    display: "flex",
-    justifyContent: "center",
-    alignItems: "center",
-  },
+export default function App() {
+  return (
+    <Router>
+      <div>
+        <ul>
+          <li>
+            <Link to="/topics">Topics</Link>
+          </li>
+        </ul>
 
-  componentWrapper: {
-    width: "600px",
-    height: "80vh",
-    border: "1px solid black",
-    display: "flex",
-    flexDirection: "column",
-  },
+        <Switch>          
+          <Route path="/topics">
+            <Topics />
+          </Route>          
+        </Switch>
+      </div>
+    </Router>
+  );
+}
 
-  h3: {
-    textAlign: "center",
-  },
-}));
-
-function Chat() {
-  const [messagesArray, setMessagesArray] = useState([]);
-
-  const classes = useStyles();
-
-  const onSendMessage = (messageText) => {
-    setMessagesArray((prev) => [
-      ...prev,
-      {
-        messageText,
-        author: "me",
-      },
-    ]);
-  };
-
-  useEffect(() => {
-    if (messagesArray.length > 0) {
-      setTimeout(() => {
-        console.log("Message was sent");
-      }, 1000);
-    }
-  }, [messagesArray]);
+function Topics() {
+  let match = useRouteMatch();
 
   return (
     <div>
-        <div className={classes.chatWrapper}>      
-        <div>
-          <h3 className={classes.h3}>name</h3>
-          <div className={classes.componentWrapper}>
-            <MessageList messagesArray={messagesArray} />
-            <MessageInput onSendMessage={onSendMessage} />
-          </div>
-        </div>
-      </div>
+      <h2>Topics</h2>
+
+      <ul>
+        <li>
+          <Link to={`${match.url}/chat1`}>First chat</Link>
+        </li>
+        <li>
+          <Link to={`${match.url}/chat2`}>
+            Second chat
+          </Link>
+        </li>
+      </ul>
+
+      {/* The Topics page has its own <Switch> with more routes
+          that build on the /topics URL path. You can think of the
+          2nd <Route> here as an "index" page for all topics, or
+          the page that is shown when no topic is selected */}
+      <Switch>
+        <Route path={`${match.path}/:topicId`}>
+          <Topic />
+        </Route>
+        <Route path={match.path}>
+          <h3>Please select a topic.</h3>
+        </Route>
+      </Switch>
     </div>
   );
 }
 
-export default Chat;
+function Topic() {
+  let { topicId } = useParams();
+  return <h3>Requested topic ID: {topicId}</h3>;
+}
